fix(invoices): generate invoice on ALL completion paths, not just office PATCH
Scheduled jobs completed in the field were never invoiced. Invoice
auto-creation lived only inside PATCH /api/jobs/:id (office modal), so
jobs marked `complete` via tech clock-out (timeclock last-tech clock-out
and tech-clock field clock_out) flipped status without ever generating
an invoice.

Extract the auto-invoice block into one idempotent helper
ensureInvoiceForCompletedJob(companyId, jobId, userId) in
lib/ensure-invoice.ts and call it from all three completion paths.

Behavior preserved verbatim from the old inline block:
- Idempotent: skips if an invoice already exists for (job_id, company_id)
- Residential (no account_id): always creates a draft
- Commercial (account_id): only when invoice_frequency = 'per_job';
  weekly/monthly still deliberately skipped (batched elsewhere)
- Discounts itemized as negative line items
- Fire-and-forget QuickBooks push, non-fatal throughout

The two clock-out call sites invoke the helper fire-and-forget so a slow
or failed invoice never blocks the clock-out response.

Forward fix only. No backfill, no commercial-billing change, no deploy.
